Created folder structure for easier readability. Made a Forms folder and moved appropriate files Made a Database folder and moved appropriate files Added Validation folder
